+++ backend/canisters/group/impl/src/lifecycle/post_upgrade.rs
@@ -1,9 +1,10 @@
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::get_upgrades_memory;
+use crate::timer_job_types::{MarkVideoCallEndedJob, TimerJob};
 use crate::{mutate_state, read_state, Data, RuntimeState};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
-use chat_events::{ChatEventInternal, MessageContentInternal};
+use chat_events::{ChatEventInternal, MessageContentInternal, Reader, OPENCHAT_BOT_USER_ID};
 use event_store_producer::{Event, EventBuilder};
 use group_canister::post_upgrade::Args;
 use ic_cdk_macros::post_upgrade;
@@ -12,9 +13,10 @@ use stable_memory::get_reader;
 use std::collections::HashMap;
 use tracing::info;
 use types::{
-    CanisterId, MessageEditedEventPayload, MessageTippedEventPayload, P2PSwapCompletedEventPayload, P2PSwapStatus,
-    ReactionAddedEventPayload, VideoCallEndedEventPayload,
+    CanisterId, MessageEditedEventPayload, MessageEventPayload, MessageTippedEventPayload, P2PSwapCompletedEventPayload,
+    P2PSwapStatus, ReactionAddedEventPayload, VideoCallEndedEventPayload,
 };
+use utils::time::HOUR_IN_MS;
 
 #[post_upgrade]
 #[trace]
@@ -78,12 +80,42 @@ fn post_upgrade(args: Args) {
     mutate_state(|state| {
         let events = extract_events(state, &token_lookup);
         state.data.event_store_client.push_many(events.into_iter(), false);
+
+        let now = state.env.now();
+        if state.data.chat.events.video_call_in_progress.timestamp < now.saturating_sub(HOUR_IN_MS) {
+            if let Some(message_id) = state
+                .data
+                .chat
+                .events
+                .video_call_in_progress
+                .value
+                .as_ref()
+                .map(|vc| vc.message_index)
+                .and_then(|index| {
+                    state
+                        .data
+                        .chat
+                        .events
+                        .main_events_reader()
+                        .message_internal(index.into())
+                        .map(|m| m.message_id)
+                })
+            {
+                state.data.timer_jobs.enqueue_job(
+                    TimerJob::MarkVideoCallEnded(MarkVideoCallEndedJob(group_canister::end_video_call::Args { message_id })),
+                    now,
+                    now,
+                );
+            }
+        }
     });
 }
 
 fn extract_events(state: &RuntimeState, token_lookup: &HashMap<CanisterId, &str>) -> Vec<Event> {
     let this_canister_id_string = state.env.canister_id().to_string();
+    let this_canister_id_str = this_canister_id_string.as_str();
     let anonymized_chat_id = state.data.chat.events.anonymized_id.clone();
+    let proposals_bot_user_id = state.data.proposals_bot_user_id;
 
     state
         .data
@@ -93,13 +125,29 @@ fn extract_events(state: &RuntimeState, token_lookup: &HashMap<CanisterId, &str>
         .flat_map(move |(e, is_thread)| {
             let mut events = Vec::new();
             if let ChatEventInternal::Message(m) = &e.event {
+                let sender_is_bot = m.sender == proposals_bot_user_id || m.sender == OPENCHAT_BOT_USER_ID;
+                events.push(
+                    EventBuilder::new("message_sent", e.timestamp)
+                        .with_user(m.sender.to_string(), true)
+                        .with_source(this_canister_id_str, true)
+                        .with_json_payload(&MessageEventPayload {
+                            message_type: m.content.message_type(),
+                            chat_type: "group".to_string(),
+                            chat_id: anonymized_chat_id.clone(),
+                            thread: is_thread,
+                            sender_is_bot,
+                            content_specific_payload: m.content.event_payload(),
+                        })
+                        .build(),
+                );
+
                 for (ledger, tips) in m.tips.iter() {
                     let token = token_lookup.get(ledger).unwrap();
                     for (user_id, amount) in tips.iter() {
                         events.push(
                             EventBuilder::new("message_tipped", e.timestamp)
                                 .with_user(user_id.to_string(), true)
-                                .with_source(this_canister_id_string.clone(), true)
+                                .with_source(this_canister_id_str, true)
                                 .with_json_payload(&MessageTippedEventPayload {
                                     message_type: m.content.message_type(),
                                     chat_type: "group".to_string(),
@@ -118,7 +166,7 @@ fn extract_events(state: &RuntimeState, token_lookup: &HashMap<CanisterId, &str>
                         events.push(
                             EventBuilder::new("reaction_added", e.timestamp)
                                 .with_user(user_id.to_string(), true)
-                                .with_source(this_canister_id_string.clone(), true)
+                                .with_source(this_canister_id_str, true)
                                 .with_json_payload(&ReactionAddedEventPayload {
                                     message_type: m.content.message_type(),
                                     chat_type: "group".to_string(),
@@ -134,7 +182,7 @@ fn extract_events(state: &RuntimeState, token_lookup: &HashMap<CanisterId, &str>
                     events.push(
                         EventBuilder::new("message_edited", e.timestamp)
                             .with_user(m.sender.to_string(), true)
-                            .with_source(this_canister_id_string.clone(), true)
+                            .with_source(this_canister_id_str, true)
                             .with_json_payload(&MessageEditedEventPayload {
                                 message_type: m.content.message_type(),
                                 chat_type: "group".to_string(),
@@ -152,7 +200,7 @@ fn extract_events(state: &RuntimeState, token_lookup: &HashMap<CanisterId, &str>
                     if let Some(ts) = video.ended {
                         events.push(
                             EventBuilder::new("video_call_ended", e.timestamp)
-                                .with_source(this_canister_id_string.clone(), true)
+                                .with_source(this_canister_id_str, true)
                                 .with_json_payload(&VideoCallEndedEventPayload {
                                     chat_type: "group".to_string(),
                                     chat_id: anonymized_chat_id.clone(),
@@ -169,7 +217,7 @@ fn extract_events(state: &RuntimeState, token_lookup: &HashMap<CanisterId, &str>
                         events.push(
                             EventBuilder::new("p2p_swap_completed", e.timestamp)
                                 .with_user(c.accepted_by.to_string(), true)
-                                .with_source(this_canister_id_string.clone(), true)
+                                .with_source(this_canister_id_str, true)
                                 .with_json_payload(&P2PSwapCompletedEventPayload {
                                     token0: swap.token0.token.token_symbol().to_string(),
                                     token0_amount: swap.token0_amount,