+++ backend/canisters/user/api/src/updates/send_message_v2.rs
@@ -2,7 +2,7 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{
     ChatId, CompletedCryptoTransaction, EventIndex, InvalidPollReason, MessageContentInitial, MessageId, MessageIndex,
-    ReplyContext, TimestampMillis, UserId,
+    Milliseconds, ReplyContext, TimestampMillis, UserId,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -14,6 +14,7 @@ pub struct Args {
     pub replies_to: Option<ReplyContext>,
     pub forwarding: bool,
     pub message_filter_failed: Option<u64>,
+    pub pin: Option<String>,
     pub correlation_id: u64,
 }
 
@@ -34,6 +35,9 @@ pub enum Response {
     P2PSwapSetUpFailed(String),
     DuplicateMessageId,
     UserSuspended,
+    PinRequired,
+    PinIncorrect(Milliseconds),
+    TooManyFailedPinAttempts(Milliseconds),
     InternalError(String),
 }