+++ backend/canisters/user/api/src/updates/send_message_with_transfer_to_group.rs
@@ -2,7 +2,7 @@ use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{
     ChatId, CompletedCryptoTransaction, Cryptocurrency, EventIndex, GroupReplyContext, MessageContentInitial, MessageId,
-    MessageIndex, TimestampMillis, User, Version,
+    MessageIndex, Milliseconds, TimestampMillis, User, Version,
 };
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -18,6 +18,7 @@ pub struct Args {
     pub correlation_id: u64,
     pub rules_accepted: Option<Version>,
     pub message_filter_failed: Option<u64>,
+    pub pin: Option<String>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -36,6 +37,9 @@ pub enum Response {
     ChatFrozen,
     RulesNotAccepted,
     Retrying(String, CompletedCryptoTransaction),
+    PinRequired,
+    PinIncorrect(Milliseconds),
+    TooManyFailedPinAttempts(Milliseconds),
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]