+++ backend/canisters/user/impl/src/updates/withdraw_crypto.rs
@@ -1,6 +1,7 @@
 use crate::crypto::process_transaction;
 use crate::guards::caller_is_owner;
-use crate::run_regular_jobs;
+use crate::model::pin_number::VerifyPinError;
+use crate::{mutate_state, run_regular_jobs};
 use canister_tracing_macros::trace;
 use ic_cdk_macros::update;
 use user_canister::withdraw_crypto_v2::{Response::*, *};
@@ -10,6 +11,14 @@ use utils::consts::MEMO_SEND;
 #[trace]
 async fn withdraw_crypto_v2(args: Args) -> Response {
     run_regular_jobs();
+
+    if let Err(error) = mutate_state(|state| state.data.pin_number.verify(args.pin.as_deref(), state.env.now())) {
+        return match error {
+            VerifyPinError::PinRequired => PinRequired,
+            VerifyPinError::PinIncorrect(delay) => PinIncorrect(delay),
+            VerifyPinError::TooManyFailedAttempted(delay) => TooManyFailedPinAttempts(delay),
+        };
+    }
 
     match process_transaction(args.withdrawal.set_memo(&MEMO_SEND)).await {
         Ok(completed_withdrawal) => Success(completed_withdrawal),